refactor(cdconc): apply chain pattern on CDCompleter & association completer level
refactor(cdcocn): reformat code
