+++ cddiff/src/main/java/de/monticore/cdconcretization/association/IAssocSideCompleter.java
@@ -1,0 +1,10 @@
+package de.monticore.cdconcretization.association;
+
+import de.monticore.cdassociation._ast.ASTCDAssocSide;
+import de.monticore.cdconcretization.CompletionException;
+
+public interface IAssocSideCompleter {
+
+  void completeAssocSide(ASTCDAssocSide concreteAssocSide, ASTCDAssocSide referenceAssocSide)
+      throws CompletionException;
+}
